refactor(tests): extract closing tests into a separate module
Move the composed workflow helper and three integration tests from the monolithic `closing.rs` into a dedicated `closing_part_01_tests.rs` file, included via a single `include!` directive. This reduces the main test file by 236 lines and groups related tests together without changing any test logic or behaviour.

Auto-committed-on: dragonfly
